Finished implementing the handicap calculator and adding the new information to the golfer update statement and the new score insert statement. Now, when a new score is added to the table, it also calculates the golfer's handicap index and updates players' averages and number of games played in the golfer table.
I polled the golfer look-up section out of the add a score function and created a new function for golfer look-up so it can also be called before adding a new golfer. Added a loop to the golfer search so that if the golfer is not found, the user is able to try entering the name again.

Linked the new lookUpGolfer function so that it is called first when adding a new golfer, updating a golfer, or adding a new score. Added an explanation that differed depending on which option was chosen, and the function also handles what to do with the golfer information differently. Currently, the update a golfer is not implemented, so it prints out a statement and returns to the main menu.