Added standerdisation of caps in the user inputs of name to be title so the serch will not fail dew to case sensativity.
Added a check for if the user entered simbales or numbers in the name and built lops around to prompt the user to enter a valed name.

Added backup searches so if an exact match is not found, it will search for just the first name, just the last name, and if neither of those matches, it will look for an approximate of eather name. Then it will prompt the user if any of those options are who they are looking for.