Configered the file set up to better work with flask.
Seperated the JavaScript out of the HTML into its own file

The program now Prints the information for the exack match golfer but if there is no exact matches it will give the user the close matches each as a button to select from then print that golfers info. Latter these print statments will be replaced with displaying the score adding form.

Added some coment tags for places where things need to be added or fixed.