Added CSS
Added stialing to the HTML page to make it feel like a golfer aplication and to look more appeling.

Added functionality so that when theuser submits the serch for the golfer it removes that section and showes the resolts box only wether exact matches or close matches. THis will be updated to show ether the add new round or option to add new golfer.