Createrd Two new pages, Add Golfer and Add Round.
Convered Intex into a landing/explunation page

Created Links for the Add Golfer and Add round Page.

Updated the Try catch Black for better Erorr handaling.

Created the start of an add round form to be expanded apon latter.

Created a function in the JS to Prefill First and Last name on the add Round form as well as a hiden Golfer ID.

Now when you enter in an exact match on the add round page it adamatickly polls up the add around form with the name and hidden ID filled in. These feilds are not editable so the user can change the nsmae that was alredy found.

Created dedicated "sections" for "MatchFound", "CloseMatch" and "noMatch" and they care hidden/revealed as nesasary bassed on the serch resolts.