Added data-mode and finished add golfer form.
I switched from using a switch statment in the Javascript to using data-mode on each of the pages to inform the JS how to act diferently depening on which page is open.

Finished setting up the form to adda new golfer and wether or not they have existing information. All the buttons for exact match, close match and no match are working corectly on both pages. When sent from, for example the add round page to add a new golfer the name that was typed in is pree filled into the new form but can not be altered. same goes for if a golfer is found on the add golfer page the page lods and prefils the add score page.